Undeprecate option -l for now..
Trying one more time implementing less pager script built-in with option -l.
What I discovered is that `less` has a 64KB buffer from stdin and that is why the script hangs after a few hours.
I thought the problem might be with the code. There was a bug in the loops which I fixed more than one year ago, but then added option -l and I thought it was the same bug that had still persisted.. 
Solution: disable `less` bufeer with option `-b-1`. Hope that really fixes it now.. Still did not test it properly but I reckon that is the fix.
That was a hard to catch introduced bug.