Many improvements in input error checking and UX.
Make user experience more consistent with what is written in the help page.
That means we should try and get current date/time if programme `date` is available in the system (which is most probable) when some of the input dates are empty, only as a fallback we should set 0001-01-01.

Some stdin input fixes. Improve check leap year option -l.

Got rid of REGEX tests completely and of the ${var^^} and  ${var,,} notations, so now this script should be able to run in Bash>3.2, however I did not test it thoroughly in such early bash versions..

No regressions should be allowed in this script. Code should be changed very slowly and with utmost care as there are too many situations it should work with and break graciously.