Changes in secondary code improve older bash compatibility
Various changes in source code to try and be more compatible with bash. It seems to be working correctly in bash2.05b. I even did some tests and the main engine seems to be working OK now.

On top of my head, we had to remove indirect variable substitution `${!var}` (we can use two eval commands for that), also make sure we don't set a negative index in any arrays (for exemple `${YEAR_MONTH_DAYS[INDEX]}`. Also, we had to remove some ternary arithmetics nested within the for-loop tests, which provided an opportunity to simplify one important for-loop, making it more readable.

We started some tests and so far it looks like the best code yet (at least no regressions! ).